pyproject.toml
[tool.poetry]
name = "pydisadm"
version = "0.1.0"
description = "A Discord bot providing ADM summaries."
authors = ["Axel Wettervik"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
schedule = "^1.2.0"
requests = "^2.30.0"
discord = "^2.2.3"
python-dotenv = "^1.0.0"
pandas = "^2.0.1"
tabulate = "^0.9.0"
matplotlib = "^3.7.1"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
python-semantic-release = "^7.33.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
version_variable = "pyproject.toml:version"